requirements.txt
dash>=2.14.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
xlrd>=2.0.0
groq>=0.4.0
python-dotenv>=1.0.0
numpy>=1.24.0